# docs/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: LexFrame Stage 8 API
  version: 0.8.0
  description: Canonical HTTP contract for the LexFrame stage 8 execution core, embedded builder, approval inbox, artifact handling, delivery and notifications.
servers:
  - url: http://localhost:3100
paths:
...
  /admin/security/alerts/{alertId}:
    patch:
      operationId: updateSecurityAlert
      tags: [Security]
      parameters:
        - in: path
          name: alertId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated security alert
          content:
            application/json:
              schema:
                type: object
  /admin/security/incidents:
    get:
      operationId: listSecurityIncidents
      tags: [Security]
      responses:
        "200":
          description: Security incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      operationId: createSecurityIncident
      tags: [Security]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Created security incident
          content:
            application/json:
              schema:
                type: object
  /admin/security/incidents/{incidentId}:
    patch:
      operationId: updateSecurityIncident
      tags: [Security]
      parameters:
        - in: path
          name: incidentId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated security incident
          content:
            application/json:
              schema:
                type: object
...
  /profiles:
    post:
      operationId: createLegalWorkProfile
      tags: [Profiles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLegalWorkProfileRequest"
      responses:
        "200":
          description: Created legal work profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalWorkProfileDetail"
  /profiles/{profileId}/draft:
    patch:
      operationId: updateLegalWorkProfileDraft
      tags: [Profiles]
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLegalWorkProfileDraftRequest"
      responses:
        "200":
          description: Updated legal work profile draft.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalWorkProfileDetail"
  /profiles/{profileId}/validate:
    post:
      operationId: validateLegalWorkProfile
      tags: [Profiles]
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile validation result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileValidationResult"
  /profiles/{profileId}/publish:
    post:
      operationId: publishLegalWorkProfile
      tags: [Profiles]
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published legal work profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegalWorkProfileDetail"
  /profiles/{profileId}/versions:
    get:
      operationId: listLegalWorkProfileVersions
      tags: [Profiles]
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile version history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LegalWorkProfileVersionSummary"
...
  /profiles/{profileId}/preview-effective:
    post:
      operationId: previewEffectiveLegalWorkProfile
      tags: [Profiles]
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewEffectiveProfileRequest"
      responses:
        "200":
          description: Effective profile preview.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EffectiveProfilePreview"
...
components:
...
  schemas:
...
    Stage7Object:
      type: object
      additionalProperties: true
    LegalWorkProfileDetail:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
...
    CreateLegalWorkProfileRequest:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    UpdateLegalWorkProfileDraftRequest:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    ProfileValidationResult:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    LegalWorkProfileVersionSummary:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
...
    PreviewEffectiveProfileRequest:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
    EffectiveProfilePreview:
      allOf:
        - $ref: "#/components/schemas/Stage7Object"
...